+++ src/main/java/io/hhplus/tdd/point/repository/PointHistoryRepository.java
@@ -1,7 +1,7 @@
 package io.hhplus.tdd.point.repository;
 
-import io.hhplus.tdd.point.PointHistory;
-import io.hhplus.tdd.point.TransactionType;
+import io.hhplus.tdd.point.domain.PointHistory;
+import io.hhplus.tdd.point.value.TransactionType;
 
 import java.util.List;
