+++ src/main/java/io/hhplus/tdd/point/repository/UserPointRepository.java
@@ -1,6 +1,6 @@
 package io.hhplus.tdd.point.repository;
 
-import io.hhplus.tdd.point.UserPoint;
+import io.hhplus.tdd.point.domain.UserPoint;
 
 public interface UserPointRepository {
     UserPoint getUserPoint(long userId);
