+++ src/test/java/io/hhplus/tdd/point/service/integration/PointServiceIntegrationTest.java
@@ -1,7 +1,7 @@
 package io.hhplus.tdd.point.service.integration;
 
-import io.hhplus.tdd.point.PointHistory;
-import io.hhplus.tdd.point.UserPoint;
+import io.hhplus.tdd.point.domain.PointHistory;
+import io.hhplus.tdd.point.domain.UserPoint;
 import io.hhplus.tdd.point.service.PointService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
