+++ src/test/java/io/hhplus/tdd/point/service/unit/PointSearchTest.java
@@ -1,6 +1,6 @@
 package io.hhplus.tdd.point.service.unit;
 
-import io.hhplus.tdd.point.UserPoint;
+import io.hhplus.tdd.point.domain.UserPoint;
 import io.hhplus.tdd.point.repository.UserPointRepository;
 import io.hhplus.tdd.point.service.PointService;
 import org.junit.jupiter.api.DisplayName;
